Fix Mode2 lux calculation, I2C error handling, and example pin comment

+++ examples/BH1750FVI_4Wire/BH1750FVI_4Wire.ino
@@ -33,7 +33,7 @@ void setup()
 
 void loop()
 {
-  uint16_t lux = LightSensor.GetLightIntensity();
+  float lux = LightSensor.GetLightIntensity();
   Serial.print("Light: ");
   Serial.println(lux);
   delay(250);
